Fx-analyzer/engine/requirements.txt
MetaTrader5>=5.0.0
pandas>=2.2.0
ta>=0.11.0
pyzmq>=25.1.2
python-dotenv>=1.0.1
numpy>=1.26.0
google-generativeai>=0.3.2
